feat: added token controller and moved all business logic in it
